+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Speed up validating annotations with user fields ([#1078](../../pull/1078))
 - Speed up validating annotation colors ([#1080](../../pull/1080))
 - Support more complex bands from the test source ([#1082](../../pull/1082))
+- Improve error thrown for invalid schema with multi source ([#1083](../../pull/1083))
 
 ### Bug Fixes
 - The cache could reuse a class inappropriately ([#1070](../../pull/1070))